requirements.txt
anyio==3.6.2
bcrypt==4.0.1
beanie==1.17.0
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.89.1
filelock==3.9.0
h11==0.14.0
identify==2.5.17
idna==3.4
install==1.3.5
lazy-model==0.0.5
motor==3.1.1
nodeenv==1.7.0
passlib==1.7.4
platformdirs==2.6.2
pre-commit==3.0.4
pyasn1==0.4.8
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.17.1
